refactor: Use u8 for consumer usage codes and add test

+++ src/key.rs
@@ -434,7 +434,7 @@ pub enum KeyUsage {
     /// Key usage code.
     Keyboard(u8),
     /// Consumer usage code.
-    Consumer(u16),
+    Consumer(u8),
     /// Custom code. (Behaviour defined by firmware implementation).
     Custom(u8),
 }
@@ -483,7 +483,7 @@ impl KeyOutput {
     }
 
     /// Constructs a [KeyOutput] from a consumer code.
-    pub fn from_consumer_code(usage_code: u16) -> Self {
+    pub fn from_consumer_code(usage_code: u8) -> Self {
         KeyOutput {
             key_code: KeyUsage::Consumer(usage_code),
             key_modifiers: KeyboardModifiers::new(),
